📦 NEW: Seperate course search
