+++ src/components/courses/index.jsx
@@ -1,4 +1,4 @@
 export { default as CourseBody } from './CourseBody'
 export { default as CourseList } from './CourseList'
-export { default as NotFoundSearch } from './NotFoundSearch'
+export { default as CourseSearch } from './CourseSearch'
 export { CourseItem } from './item'
