+++ src/components/shared/index.jsx
@@ -4,6 +4,7 @@ export { default as Divider } from './Divider'
 export { default as InputRounded } from './InputRounded'
 export { InputSquared, TextAreaSquared } from './InputSquared'
 export { default as LoaderAnimation } from './LoaderAnimation'
+export { default as NotFoundSearch } from './NotFoundSearch'
 export { PageHeading, PageTitle } from './Page'
 export { default as ProfileImg } from './ProfileImg'
 export { default as ReadMore } from './ReadMore'
